+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ---
-order: 4
+order: 3
 chinese: 更新日志
 ---